requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
twilio==9.3.1
pydantic==2.6.1
openai==1.58.1
python-dotenv==1.0.1
